requirements.txt
google-cloud-aiplatform
langchain-google-vertexai
langchain
python-dotenv